Finish test
